requirements.txt
django
gunicorn
django-heroku
requests
python-binance
django-fsm
djangorestframework
rq